requirements.txt
aerich==0.7.2
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.6.2.post1
asyncpg==0.30.0
bcrypt==4.2.1
click==8.1.7
dictdiffer==0.9.0
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.5
h11==0.14.0
idna==3.10
iso8601==2.1.0
Jinja2==3.1.4
MarkupSafe==3.0.2
mypy==1.13.0
mypy-extensions==1.0.0
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.10.0
pypika-tortoise==0.2.2
python-dotenv==1.0.1
pytz==2024.2
sniffio==1.3.1
starlette==0.41.3
tomlkit==0.13.2
tortoise-orm==0.22.0
typing_extensions==4.12.2
uvicorn==0.32.0
